Move vehicle parking updaters into sandbox
