+++ docs/SandboxExtension.md
@@ -22,6 +22,7 @@ nearby stops generated by routing via OSM data.
 - [Mapbox Vector Tiles API](sandbox/MapboxVectorTilesApi.md) - Mapbox Vector Tiles API
 - [Flex Routing](sandbox/Flex.md) - Flexible transit routing for GTFS and Netex data sources
 - [Park and Ride API](sandbox/ParkAndRideApi.md) - Park and Ride API
+- [Vehicle Parking](sandbox/VehicleParking.md) - Vehicle Parking updaters
 
 
 ## Terminology
