+++ docs/sandbox/VehicleParking.md
@@ -1,0 +1,13 @@
+# Vehicle Parking Updaters - OTP Sandbox Extension
+
+## Contact Info
+TODO
+
+## Changelog
+TODO
+
+## Documentation
+TODO
+
+### Configuration
+TODO
