+++ mkdocs.yml
@@ -53,3 +53,4 @@ pages:
     - Interactive OTP Launcher: 'sandbox/InteractiveOtpMain.md'
     - Sandbox Extension Example: 'sandbox/Examples.md'
     - Park and Ride API: 'sandbox/ParkAndRideApi.md'
+    - Smoove Bike Rental Updator Support: 'sandbox/VehicleParking.md'
