+++ src/ext-test/resources/vehicleparking/kml/NSFietsenstallingen.kml
