+++ src/ext-test/resources/vehicleparking/parkapi/parkapi-reutlingen.json
